+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [UNRELEASED]
 
+### Added
+
+- Make the registration params and profile params configurable.
+
 ## [0.2.0] - 2024-05-02
 
 ### Added